Make the movement behaviour consistent with the original game.
The previous behaviour was wrong since it was creating a new
"block" even no changes was made in the table. This way if
the player gets "stucked" he could only keep trying the same
movement direction and a new block will be created every time.

In the original game a new block only is spawned if the "table"
state gets changes, i.e the player must do a valid move to a new
block get spawned.

We accomplish that by copying the current TABLE that is passed
to the key() function. This way we can compare the previous and
now "current" state of the table to check if a move was did.
Only then, we spawn a new block.

Make a little bit of refactoring on the main loop, since previously
we have too much code duplicated.

Remove the inconsistency of tabs vs spaces.